+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.17.4
 
+### Improvements
+- Better control dtype on multi sources ([#993](../../pull/993))
+
 ### Bug Fixes
 - Use open.read rather than download to access files in Girder ([#989](../../pull/989))
 - Fix nd2 source scale ([#990](../../pull/990))