+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.17.4
 
+### Improvements
+- Better control dtype on multi sources ([#993](../../pull/993))
+- Don't use dask threads when using nd2 to fetch tiles ([#994](../../pull/994))
+- Set mime type for imported girder files ([#995](../../pull/995))
+
 ### Bug Fixes
 - Use open.read rather than download to access files in Girder ([#989](../../pull/989))
 - Fix nd2 source scale ([#990](../../pull/990))